requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
sentence-transformers==3.0.1
pypdf==4.3.1
httpx==0.27.0
Jinja2==3.1.4
numpy==2.1.3